Rename DELETE /api/v1/annotations/ to POST /api/v1/retire_annotations
This is a mitigation to the issue raised in LEARNER-6532 where
apparently buggy annotator.js code is calling the DELETE endpoint
without providing an annotation id.  Our annotator-compatible API
deviates from the documented annotator API in such a way that NOT
providing the annotation id implies delete (retire) all annotations.

The correct fix should be to additionally protect this API endpoint by
only authorizing the LMS, but renaming the endpoint is a sufficient
mitigation since annotator.js is far less likely to accidentally call
it.