feat: meilisearch integration (#453)
* chore: Update catalog.yaml file for release data

* chore: Delete openedx.yaml file

* fix: refactor workflows

* chore: Upgrade Python requirements (#450)

* fix: path import is fixed
After upgrading path-py from 9.1 to 12.5.0 they upgrade the import as well from p to capital P as import path from Path

* fix: enable quality checks

* chore: Upgrade Python requirements (#452)

* chore: Upgrade Python requirements

* fix: trivy action failure

---------

Co-authored-by: Irtaza Akram <irtaza.akram@arbisoft.com>

* feat: meilisearch backend for notes search (#444)

* feat: introduce "make compile-requirement" target

This is convenient to compile dependencies without upgrading them.

* chore: simplify tox/make test commands

This makes it possible to run the make commands directly without going
through tox (though tox targets keep working, of course).

* chore: more convenient unit test running

Previously, it was not possible to run unit tests locally without
manually creating mysql & elasticsearch containers. Here, we create a
`make pytest` target that automatically starts the required containers.

* chore: refactor views for better mysql/es separation

Instead of checking a boolean flag in multiple different places, we use
class inheritance. This makes it possible to later override the view and
implement our own using a different search backend, such as Meilisearch.

* feat: meilisearch backend for notes search

This is a very simple and basic backend. It is based on Django signals,
just like the Elasticsearch backend. But it is much simpler, in the
sense that there are just two signals: one for saving documents and one
for deletion.

This backend is limited, in the sense that it does not support
highlighting -- but that's probably not such a big deal.

To start using this backend, define the following settings:

	ES_DISABLED = True
	MEILISEARCH_ENABLED = True
	MEILISEARCH_URL = "http://meilisearch:7700"
	MEILISEARCH_API_KEY = "s3cr3t"
	MEILISEARCH_INDEX = "tutor_student_notes"

---------

Co-authored-by: salman2013 <salman.nawaz@arbisoft.com>
Co-authored-by: Irtaza Akram <irtaza.akram@arbisoft.com>
Co-authored-by: edX requirements bot <49161187+edx-requirements-bot@users.noreply.github.com>
Co-authored-by: Muhammad Arslan <arslan.abdulrauf@arbisoft.com>
Co-authored-by: Irtaza Akram <51848298+irtazaakram@users.noreply.github.com>